[e2e] Extract eviction strategy related tests in its own file
Migration test file is growing and now it's hard to read. Also,
different contexts are overlapping causing several bugs recently.

Also:
- collapsed the BeforeEach functions
- exported utility functions needed by eviction and migration file
