+++ tests/migration/BUILD.bazel
@@ -3,6 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = [
+        "eviction_strategy.go",
         "framework.go",
         "migration.go",
         "migration_policy.go",
